refactor(19240): refactor pipeline checks for usage with both data and bahavior policies
