+++ hivemq-edge/src/frontend/src/extensions/datahub/designer/data_policy/DataPolicyNode.utils.ts
@@ -1,8 +1,6 @@
 import { getIncomers, getOutgoers, Node } from 'reactflow'
-import { PolicyOperation } from '@/api/__generated__'
 
 import { DataHubNodeType, DataPolicyData, DryRunResults, TopicFilterData, WorkspaceState } from '@datahub/types.ts'
-import { checkValidityTransformFunction } from '@datahub/designer/operation/OperationNode.utils.ts'
 import { PolicyCheckErrors } from '@datahub/designer/validation.errors.ts'
 
 export function getSubFlow(source: Node, acc: Node[], store: WorkspaceState, only = false) {
@@ -72,57 +70,4 @@ export function checkValidityFilter(
     // TODO[XXXX] Multiple topics?
     data: incomers[0].data.topics[0],
   }
-}
-
-export function checkValidityPipeline(
-  dataPolicyNode: Node<DataPolicyData>,
-  handle: DataPolicyData.Handle,
-  store: WorkspaceState
-): DryRunResults<PolicyOperation>[] {
-  const { nodes, edges } = store
-
-  const getNextNode = (node: Node | undefined): Node | undefined => {
-    if (node) {
-      const outEdge = edges.find((edge) => edge.source === node.id)
-      if (outEdge) {
-        const nextNode = nodes.find((node) => node.id === outEdge.target)
-        if (nextNode) return nextNode
-      }
-    }
-    return undefined
-  }
-
-  const [outboundEdge] = edges.filter((edge) => edge.source === dataPolicyNode.id && edge.sourceHandle === handle)
-  if (!outboundEdge) {
-    return []
-  }
-
-  const pipeline: Node[] = []
-  let nextNode = nodes.find((node) => node.id === outboundEdge.target)
-  while (nextNode) {
-    pipeline.push(nextNode)
-    nextNode = getNextNode(nextNode)
-  }
-
-  return pipeline.map((node) => {
-    if (!node.data.functionId) {
-      return {
-        node: node,
-        error: PolicyCheckErrors.notConfigured(node, 'functionId'),
-      }
-    }
-
-    if (node.data.functionId === 'DataHub.transform') {
-      return checkValidityTransformFunction(node, store)
-    }
-
-    // TODO[NVL] Serialisers need to be dealt with
-
-    const operation: PolicyOperation = {
-      functionId: node.data.functionId,
-      arguments: node.data.formData,
-      id: node.id,
-    }
-    return { node: node, data: operation }
-  })
 }
